refactor(breakpoints): create different files for production and development environment
