+++ breakpoints/index.js
@@ -1,0 +1,5 @@
+if (process.env.NODE_ENV === 'production') {
+  module.exports = require('./breakpoints.js');
+} else {
+  module.exports = require('./breakpoints.dev.js');
+}
