--- core/index.js
@@ -1,3 +1,0 @@
-const { createBreakpoints } = require('./breakpoints');
-
-exports.createBreakpoints = createBreakpoints;
